+++ M02S11/Exercicio/quiz/src/main/java/tech/devinhouse/quiz/service/QuizService.java
@@ -1,8 +1,14 @@
 package tech.devinhouse.quiz.service;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import tech.devinhouse.quiz.mapper.QuizMapper;
 import tech.devinhouse.quiz.repository.model.Quiz;
 import tech.devinhouse.quiz.repository.QuizRepository;
+import tech.devinhouse.quiz.repository.model.dto.QuizRequest;
 
 import java.util.List;
 import java.util.Optional;
@@ -16,20 +22,59 @@ public class QuizService {
         this.quizRepository = quizRepository;
     }
 
-    public List<Quiz> getAll(){
-        return this.quizRepository.findAll();
+    public ResponseEntity<Object> getAll(){
+        List<Quiz> quizzes = this.quizRepository.findAll();
+
+        if (quizzes.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.OK).body("Nenhum Quiz Cadastrado");
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(quizzes);
     }
 
-    public Optional<Quiz> getById(Long id){
-        return this.quizRepository.findById(id);
+    public ResponseEntity<Object> getById(Long id){
+
+        Optional<Quiz> quiz = this.quizRepository.findById(id);
+
+        if (quiz.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Nenhum Quiz Cadastrado");
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(quiz);
     }
 
-    public Quiz saveQuiz(Quiz quiz){
-        return this.quizRepository.save(quiz);
+    public ResponseEntity<Object> saveQuiz(QuizRequest quizrRequest){
+
+        Quiz quiz = QuizMapper.INSTANCE.reqToQuiz(quizrRequest);
+        Quiz newQuiz = this.quizRepository.save(quiz);
+
+        return ResponseEntity.status(HttpStatus.CREATED).body(newQuiz);
+
     }
 
-    public void deleteQuiz(Long id){
+    public ResponseEntity<Object> updatePergunta(Long id, QuizRequest quizRequest) {
+
+        Optional<Quiz> opQuiz = this.quizRepository.findById(id);
+
+        if (opQuiz.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Nenhum Quiz Cadastrado");
+        }
+
+
+        Quiz quiz = QuizMapper.INSTANCE.reqToQuiz(quizRequest);
+        quiz.setId(id);
+        Quiz newQuiz = this.quizRepository.save(quiz);
+        return ResponseEntity.status(HttpStatus.CREATED).body(newQuiz);
+    }
+
+    public ResponseEntity<Object> deleteQuiz(Long id){
+        Optional<Quiz> quiz = this.quizRepository.findById(id);
+
+        if (quiz.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Nenhum Quiz Cadastrado");
+        }
         this.quizRepository.deleteById(id);
+        return ResponseEntity.status(HttpStatus.OK).body(null);
+
+
     }
